requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
zappa
python-dotenv
Flask-Bcrypt
Flask-JWT-Extended
flask-cors
Flask-Migrate
pytz
apscheduler